Flow upgrade to 0.146
This upgrade made more expressions invalidate refinements. In some
places this lead to a large number of suppressions that I automatically
suppressed and should be followed up on when the code is touched.
I think most of them might require either manual annotations or moving
a value into a const to allow refinement.

ghstack-source-id: a45b40abf03474b691a2dc18a4f1f3b06cfbdffd
Pull Request resolved: https://github.com/facebook/react/pull/25410